Fixed file reading and writing
Optimized file writing in CharacterCreatorActivity to minimize the
amount of lines and wrote it as a parcelable rather than a plain
text file. Created CharacterDetailActivity and open the selected
CharacterSheetData file upon creation of the activity.

When the user selects a character from the list, it will now open
to a character sheet with the character's name listed.