Fixed skill modifier calculations
Changed the calculation of trained skills to include the proficiency
modifier. CharacterDetailAcitivity also properly displays a character's
proficiency bonus and corresponding skill modifiers.

When a value has been entered for the proficiency modifier, it will be
applied to all trained skills.  When the character sheet is viewed, it
will show the proper proficiency modifier and skill modifiers.